Update components.yaml to fix wrong answer and wrong link documentation
The object variables doesn't exist and can't be used in the ExpressionLanguage for matching expressions. It should be params variable because the variables exists are context, params and request and the existing methods are env() and service()

Besides, the Symfony 7 Certification exam only includes questions about Symfony 7.0 and not about Symfony 7.1, 7.2, 7.3 and 7.4 versions.